Added environment based application loading.
Server access to dev environments was previously blocked on every IP except 127.0.0.1
for the dev environment, and also only switched based on .htaccess file or direct url access.
This issue was fixed by detecting the APPLICATION_ENV, and changing which app.php
file was loaded. In addition, app_dev.php is explicitly blocked unless the dev
environment is set for the application.

Conflicts:
	web/app_dev.php