Fixing cache deletion problem
A problem exists in symfony that the Session cache files do not
respect the umask settings for the project. This patch fixes that
by removing the whole directory in one command, as opposed to phing's
command that removes individual files.
This should be improved in the future to include Windows systems