Adding RequireJS Skeleton
Added in support for RequireJS into the project

The RequireJSBundle was added to the project along with r.js for optimization
during production runs. It should be noted that a node exectuable is required
for this to run properly. The r.js dependency is included as a raw git repo
due to the fact that no maintained composer package exists to support it.